251118 Aula 15 Fim de Aula

+++ ProgPY_Não_POO/Aula11/pessoasIdadeAlturaST.py
@@ -12,13 +12,13 @@ nomes_menor16 = []
 soma_alt=0
 qtde_menor16=0
 while True:
-    n=st.number_input('Digite o nr. de pessoas (máximo de 10 pessoas): ')
+    n=int(st.number_input('Digite o nr. de pessoas (máximo de 10 pessoas): '))
     if n>MAXIMO:
         st.error('Excedeu o nr. máximo de 10 pessoas.')
         continue
     else:
         break
-qtde=n+1
+qtde=int(n+1)
 for i in range(1,qtde):
     nomes.append=st.text_input(f'Digite o nome da {i}ª pessoa: ')
     idades.append=st.number_input(f'Idade (anos inteiro) de {nomes[i-1]}: ')
@@ -29,8 +29,8 @@ for i in range(1,qtde):
         qtde_menor16+=1
 
 # Processamento
-media_altura=soma_alt/n
-perc_menor16=(qtde_menor16/n)*100
+media_altura=float(soma_alt/n)
+perc_menor16=float((qtde_menor16/n)*100)
 
 # Saída de dados
 st.success(f'A média da altura de todas as {n} pessoas é: {media_altura:.2f}')
